Add files via upload
the customer needed to collate ArcGIS measurement across multiple test executions.  They then are looking at 90th percentile.  
The 90th percentile refers to the value below which 90% of the data in a dataset falls.  This PowerShell script will take an array of testrunids and collect all of the application measurements.  It will then take those measurements and calculate the average duration value when the application ID and the measurement id are the same between multiple testrunids.  Right now it only calculates the average but the math can be adjusted just needed to get the logic in place.